Rename too-many-positional(-arguments)
