+++ crates/ruff_linter/resources/test/fixtures/pylint/too_many_positional_arguments.py
