+++ crates/ruff_linter/src/rules/pylint/rules/mod.rs
@@ -79,7 +79,7 @@ pub(crate) use too_many_boolean_expressions::*;
 pub(crate) use too_many_branches::*;
 pub(crate) use too_many_locals::*;
 pub(crate) use too_many_nested_blocks::*;
-pub(crate) use too_many_positional::*;
+pub(crate) use too_many_positional_arguments::*;
 pub(crate) use too_many_public_methods::*;
 pub(crate) use too_many_return_statements::*;
 pub(crate) use too_many_statements::*;
@@ -182,7 +182,7 @@ mod too_many_boolean_expressions;
 mod too_many_branches;
 mod too_many_locals;
 mod too_many_nested_blocks;
-mod too_many_positional;
+mod too_many_positional_arguments;
 mod too_many_public_methods;
 mod too_many_return_statements;
 mod too_many_statements;
